route 중복 제거 및 코드정리
