+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@prisma/client": "^6.8.2",
-    "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
     "debug": "~2.6.9",
     "dotenv": "^16.5.0",
